fix(baseN, hex, sha256): added str_size argument

+++ include/base/hash/sha256.hpp
@@ -7,6 +7,6 @@
 #define SHA256_DIGEST_LENGTH 32
 
 namespace hash {
-	void sha256(const uint8_t *data, uint64_t data_size, uint8_t *hash) noexcept;
+	void sha256(const uint8_t *data, size_t data_size, uint8_t *hash) noexcept;
 	std::vector<uint8_t> sha256(std::span<const uint8_t> data) noexcept;
 }
